@@ -6,7 +6,6 @@ import dev.brighten.antivpn.AntiVPN;
 import dev.brighten.antivpn.api.APIPlayer;
 import dev.brighten.antivpn.api.VPNExecutor;
 import dev.brighten.antivpn.web.objects.VPNResponse;
-import net.md_5.bungee.BungeeCord;
 import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.chat.TextComponent;
 import net.md_5.bungee.api.event.PlayerDisconnectEvent;
@@ -37,17 +36,8 @@ public class BungeeListener extends VPNExecutor implements Listener {
     }
 
     @Override
-    public void onShutdown() {
-        if(cacheResetTask != null) {
-            cacheResetTask.cancel();
-            cacheResetTask = null;
-        }
-        BungeePlugin.pluginInstance.getProxy().getPluginManager().unregisterListener(this);
-    }
-
-    @Override
     public void log(Level level, String log, Object... objects) {
-        BungeeCord.getInstance().getLogger().log(Level.INFO, String.format(log, objects));
+        BungeePlugin.pluginInstance.getProxy().getLogger().log(Level.INFO, String.format(log, objects));
     }
 
     @Override
@@ -56,18 +46,18 @@ public class BungeeListener extends VPNExecutor implements Listener {
     }
 
     @Override
-    public void logException(String message, Exception ex) {
-        BungeeCord.getInstance().getLogger().log(Level.SEVERE, message, ex);
+    public void logException(String message, Throwable ex) {
+        BungeePlugin.pluginInstance.getProxy().getLogger().log(Level.SEVERE, message, ex);
     }
 
     @Override
     public void disablePlugin() {
-        BungeeCord.getInstance().getPluginManager().unregisterListeners(BungeePlugin.pluginInstance);
+        BungeePlugin.pluginInstance.getProxy().getPluginManager().unregisterListeners(BungeePlugin.pluginInstance);
         if(cacheResetTask != null) {
             cacheResetTask.cancel();
             cacheResetTask = null;
         }
-        BungeeCord.getInstance().getPluginManager().unregisterCommands(BungeePlugin.pluginInstance);
+        BungeePlugin.pluginInstance.getProxy().getPluginManager().unregisterCommands(BungeePlugin.pluginInstance);
         BungeePlugin.pluginInstance.onDisable();
     }
 
@@ -79,7 +69,7 @@ public class BungeeListener extends VPNExecutor implements Listener {
 
         if(cached != null && cached.isProxy()) {
             event.setCancelled(true);
-            event.setCancelReason(TextComponent.fromLegacyText(ChatColor
+            event.setReason(TextComponent.fromLegacy(ChatColor
                     .translateAlternateColorCodes('&',
                             AntiVPN.getInstance().getVpnConfig().getKickString())));
             AntiVPN.getInstance().getExecutor().log(Level.INFO,
@@ -94,92 +84,92 @@ public class BungeeListener extends VPNExecutor implements Listener {
                 || AntiVPN.getInstance().getVpnConfig().getPrefixWhitelists().stream()
                 .anyMatch(prefix -> event.getPlayer().getName().startsWith(prefix))) return;
 
-        checkIp(event.getPlayer().getAddress().getAddress().getHostAddress(),
-                AntiVPN.getInstance().getVpnConfig().cachedResults(), result -> {
-            if(result.isSuccess()) {
-                //If the player is whitelisted, we don't want to kick them
-                if(AntiVPN.getInstance().getExecutor().isWhitelisted(event.getPlayer().getUniqueId())) {
-                    AntiVPN.getInstance().getExecutor().log("UUID is whitelisted: %s",
-                            event.getPlayer().getUniqueId().toString());
-                    return;
-                }
+        String address = event.getPlayer().getSocketAddress().toString();
 
-                //If the IP is whitelisted, we don't want to kick them
-                if(AntiVPN.getInstance().getExecutor().isWhitelisted(event.getPlayer().getAddress().getAddress()
-                        .getHostAddress())) {
-                    AntiVPN.getInstance().getExecutor().log("IP is whitelisted: %s",
-                            event.getPlayer().getAddress().getAddress().getHostAddress());
-                    return;
-                }
+        if(AntiVPN.getInstance().getExecutor().isWhitelisted(event.getPlayer().getUniqueId())) {
+            AntiVPN.getInstance().getExecutor().log("UUID is whitelisted: %s",
+                    event.getPlayer().getUniqueId().toString());
+            return;
+        }
 
-                responseCache.put(event.getPlayer().getUniqueId(), result);
+        //If the IP is whitelisted, we don't want to kick them
+        if(AntiVPN.getInstance().getExecutor().isWhitelisted(address)) {
+            AntiVPN.getInstance().getExecutor().log("IP is whitelisted: %s", address);
+            return;
+        }
 
-                if(!AntiVPN.getInstance().getVpnConfig().countryList().isEmpty()
-                        // This bit of code will decide whether or not to kick the player
-                        // If it contains the code and it is set to whitelist, it will not kick as they are equal
-                        // and vise versa. However, if the contains does not match the state, it will kick.
-                        && AntiVPN.getInstance().getVpnConfig().countryList()
-                        .contains(result.getCountryCode()) != AntiVPN.getInstance().getVpnConfig().whitelistCountries()) {
-                    //Using our built in kicking system if no commands are configured
-                    if(AntiVPN.getInstance().getVpnConfig().countryKickCommands().isEmpty()) {
-                        final String kickReason = AntiVPN.getInstance().getVpnConfig()
-                                .countryVanillaKickReason();
-                        // Kicking our player
-                        event.getPlayer().disconnect(TextComponent.fromLegacyText(ChatColor
-                                .translateAlternateColorCodes('&',
-                                        kickReason
+        checkIp(address)
+                .thenAccept(result -> {
+                    if(result.isSuccess()) {
+                        //If the player is whitelisted, we don't want to kick them
+                        responseCache.put(event.getPlayer().getUniqueId(), result);
+                        if(!AntiVPN.getInstance().getVpnConfig().countryList().isEmpty()
+                                // This bit of code will decide whether or not to kick the player
+                                // If it contains the code and it is set to whitelist, it will not kick as they are equal
+                                // and vise versa. However, if the contains does not match the state, it will kick.
+                                && AntiVPN.getInstance().getVpnConfig().countryList()
+                                .contains(result.getCountryCode()) != AntiVPN.getInstance().getVpnConfig().whitelistCountries()) {
+                            //Using our built in kicking system if no commands are configured
+                            if(AntiVPN.getInstance().getVpnConfig().countryKickCommands().isEmpty()) {
+                                final String kickReason = AntiVPN.getInstance().getVpnConfig()
+                                        .countryVanillaKickReason();
+                                // Kicking our player
+                                event.getPlayer().disconnect(TextComponent.fromLegacy(ChatColor
+                                        .translateAlternateColorCodes('&',
+                                                kickReason
+                                                        .replace("%player%", event.getPlayer().getName())
+                                                        .replace("%country%", result.getCountryName())
+                                                        .replace("%code%", result.getCountryCode()))));
+                            } else {
+                                for (String cmd : AntiVPN.getInstance().getVpnConfig().countryKickCommands()) {
+                                    final String formattedCommand = ChatColor.translateAlternateColorCodes('&',
+                                            cmd.replace("%player%", event.getPlayer().getName())
+                                                    .replace("%country%", result.getCountryName())
+                                                    .replace("%code%", result.getCountryCode()));
+
+                                    // Runs our command from console
+                                    BungeePlugin.pluginInstance.getProxy().getPluginManager().dispatchCommand(
+                                            BungeePlugin.pluginInstance.getProxy().getConsole(), formattedCommand);
+                                }
+                            }
+                        } else if(result.isProxy()) {
+                            if(AntiVPN.getInstance().getVpnConfig().kickPlayersOnDetect())
+                                event.getPlayer().disconnect(TextComponent.fromLegacy(ChatColor
+                                        .translateAlternateColorCodes('&',
+                                                AntiVPN.getInstance().getVpnConfig().getKickString())));
+                            BungeePlugin.pluginInstance.getProxy().getLogger().info(event.getPlayer().getName()
+                                    + " joined on a VPN/Proxy (" + result.getMethod() + ")");
+
+                            if(AntiVPN.getInstance().getVpnConfig().alertToStaff()) //Ensuring the user wishes to alert to staff
+                                AntiVPN.getInstance().getPlayerExecutor().getOnlinePlayers().stream()
+                                        .filter(APIPlayer::isAlertsEnabled)
+                                        .forEach(pl -> pl.sendMessage(AntiVPN.getInstance().getVpnConfig()
+                                                .alertMessage()
                                                 .replace("%player%", event.getPlayer().getName())
+                                                .replace("%reason%", result.getMethod())
                                                 .replace("%country%", result.getCountryName())
-                                                .replace("%code%", result.getCountryCode()))));
+                                                .replace("%city%", result.getCity())));
+
+                            //In case the user wants to run their own commands instead of using the built in kicking
+                            if(AntiVPN.getInstance().getVpnConfig().runCommands()) {
+                                for (String command : AntiVPN.getInstance().getVpnConfig().commands()) {
+                                    BungeePlugin.pluginInstance.getProxy().getPluginManager()
+                                            .dispatchCommand(BungeePlugin.pluginInstance.getProxy().getConsole(),
+                                                    ChatColor.translateAlternateColorCodes('&',
+                                                            command.replace("%player%", event.getPlayer().getName())));
+                                }
+                            }
+                            AntiVPN.getInstance().detections++;
+                        }
+
                     } else {
-                        for (String cmd : AntiVPN.getInstance().getVpnConfig().countryKickCommands()) {
-                            final String formattedCommand = ChatColor.translateAlternateColorCodes('&',
-                                    cmd.replace("%player%", event.getPlayer().getName())
-                                            .replace("%country%", result.getCountryName())
-                                            .replace("%code%", result.getCountryCode()));
-
-                            // Runs our command from console
-                            BungeeCord.getInstance().getPluginManager().dispatchCommand(
-                                    BungeeCord.getInstance().getConsole(), formattedCommand);
-                        }
+                        BungeePlugin.pluginInstance.getProxy().getLogger()
+                                .log(Level.WARNING,
+                                        "The API query was not a success! " +
+                                                "You may need to upgrade your license on https://funkemunky.cc/shop");
                     }
-                } else if(result.isProxy()) {
-                    if(AntiVPN.getInstance().getVpnConfig().kickPlayersOnDetect())
-                        event.getPlayer().disconnect(TextComponent.fromLegacyText(ChatColor
-                                .translateAlternateColorCodes('&',
-                                        AntiVPN.getInstance().getVpnConfig().getKickString())));
-                    BungeeCord.getInstance().getLogger().info(event.getPlayer().getName()
-                            + " joined on a VPN/Proxy (" + result.getMethod() + ")");
-
-                    if(AntiVPN.getInstance().getVpnConfig().alertToStaff()) //Ensuring the user wishes to alert to staff
-                        AntiVPN.getInstance().getPlayerExecutor().getOnlinePlayers().stream()
-                                .filter(APIPlayer::isAlertsEnabled)
-                                .forEach(pl -> pl.sendMessage(AntiVPN.getInstance().getVpnConfig().alertMessage()
-                                        .replace("%player%", event.getPlayer().getName())
-                                        .replace("%reason%", result.getMethod())
-                                        .replace("%country%", result.getCountryName())
-                                        .replace("%city%", result.getCity())));
-
-                    //In case the user wants to run their own commands instead of using the built in kicking
-                    if(AntiVPN.getInstance().getVpnConfig().runCommands()) {
-                        for (String command : AntiVPN.getInstance().getVpnConfig().commands()) {
-                            BungeeCord.getInstance().getPluginManager()
-                                    .dispatchCommand(BungeeCord.getInstance().getConsole(),
-                                            ChatColor.translateAlternateColorCodes('&',
-                                                    command.replace("%player%", event.getPlayer().getName())));
-                        }
-                    }
-                    AntiVPN.getInstance().detections++;
-                }
-
-            } else {
-                BungeeCord.getInstance().getLogger()
-                        .log(Level.WARNING,
-                                "The API query was not a success! " +
-                                        "You may need to upgrade your license on https://funkemunky.cc/shop");
-            }
-            AntiVPN.getInstance().checked++;
-        });
+                    AntiVPN.getInstance().checked++;
+                });
     }
 
     @EventHandler
