@@ -8,9 +8,9 @@ import lombok.Getter;
 
 import java.io.IOException;
 import java.util.*;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
-import java.util.function.Consumer;
 import java.util.logging.Level;
 
 public abstract class VPNExecutor {
@@ -22,15 +22,13 @@ public abstract class VPNExecutor {
     private final Set<String> whitelistedIps = Collections.synchronizedSet(new HashSet<>());
     public abstract void registerListeners();
 
-    public abstract void onShutdown();
-
     public abstract void log(Level level, String log, Object... objects);
 
     public abstract void log(String log, Object... objects);
 
-    public abstract void logException(String message, Exception ex);
+    public abstract void logException(String message, Throwable ex);
 
-    public void logException(Exception ex) {
+    public void logException(Throwable ex) {
         logException("An exception occurred: " + ex.getMessage(), ex);
     }
 
@@ -48,33 +46,31 @@ public abstract class VPNExecutor {
         return whitelistedIps.contains(ip);
     }
 
-    public void checkIp(String ip, boolean cachedResults, Consumer<VPNResponse> result) {
-        threadExecutor.execute(() -> result.accept(checkIp(ip)));
-    }
+    public CompletableFuture<VPNResponse> checkIp(String ip) {
+        return CompletableFuture.supplyAsync(() -> {
+            Optional<VPNResponse> cachedRes = AntiVPN.getInstance().getDatabase().getStoredResponse(ip);
 
-    public VPNResponse checkIp(String ip) {
-        Optional<VPNResponse> cachedRes = AntiVPN.getInstance().getDatabase().getStoredResponse(ip);
-
-        if(cachedRes.isPresent()) {
-            return cachedRes.get();
-        }
-        else {
-            try {
-                VPNResponse response = FunkemunkyAPI
-                        .getVPNResponse(ip, AntiVPN.getInstance().getVpnConfig().getLicense(), true);
-
-                if (response.isSuccess()) {
-                    AntiVPN.getInstance().getDatabase().cacheResponse(response);
-                } else {
-                    log("Query to VPN API failed! Reason: " + response.getFailureReason());
-                }
-
-                return response;
-            } catch (JSONException | IOException e) {
-                log("Query to VPN API failed! Reason: " + e.getMessage());
-                return VPNResponse.FAILED_RESPONSE;
+            if(cachedRes.isPresent()) {
+                return cachedRes.get();
             }
-        }
+            else {
+                try {
+                    VPNResponse response = FunkemunkyAPI
+                            .getVPNResponse(ip, AntiVPN.getInstance().getVpnConfig().getLicense(), true);
+
+                    if (response.isSuccess()) {
+                        AntiVPN.getInstance().getDatabase().cacheResponse(response);
+                    } else {
+                        log("Query to VPN API failed! Reason: " + response.getFailureReason());
+                    }
+
+                    return response;
+                } catch (JSONException | IOException e) {
+                    log("Query to VPN API failed! Reason: " + e.getMessage());
+                    return VPNResponse.FAILED_RESPONSE;
+                }
+            }
+        }, threadExecutor);
     }
 
     public abstract void disablePlugin();
